handled unavailable browser installation
